instructions separated from getters
