+++ src/data/access.rs
@@ -1,0 +1,15 @@
+use super::INSTRUCTIONS;
+use crate::model::{Funct3, Funct7, Mnemonic, Opcode};
+
+pub fn get_mnemonic(
+    opcode: Opcode,
+    funct3: Option<Funct3>,
+    funct7: Option<Funct7>,
+) -> Option<Mnemonic> {
+    let op: u16 = (u8::from(opcode) & 0b1111100).into();
+    let f3: u16 = funct3.map_or(0, |val| u16::from(u8::from(val)));
+    let f7: u16 = funct7.map_or(0, |val| u16::from(u8::from(val)));
+    let code: u16 = (op << 8) | (f3 << 7) | f7;
+
+    INSTRUCTIONS.get(&code).map(|res| res.0.into())
+}
